Move ArrayResponseType to message/responsetypes
Move ArrayResponseType to message/responsetypes. In doing so, we reflect
 the folder structure for the other ResponseType implementation in Axon
 Framework

#338
